TP3/pyproject.toml
[project]
name = "tp3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.5.2",
    "kafka-python>=2.2.15",
    "numpy>=2.3.3",
    "scikit-learn>=1.7.2",
]
